Add Extra Pages for Details
Extra pages will be used for details on each matter described in the
first page, as well as any other items excluded from the first page.

Also adds the `fancyhdr` package and a new header-footer format to
distinguish between the first page (which has no header-footers) and
other pages (which have headers and footers).